+++ Cargo.toml
@@ -56,7 +56,7 @@ sha2 = "0.9.4"
 smallvec = "1.6.1"
 structopt = { version = "0.3.21", optional = true }
 thiserror = "1.0.24"
-tokio = { version = "1.5.0", features = ["full"], optional = true }
+tokio = { version = "1.9.0", features = ["full"], optional = true }
 tracing = "0.1.26"
 tracing-error = "0.1.2"
 tracing-futures = "0.2.5"
@@ -64,3 +64,6 @@ tracing-subscriber = { version = "0.2.18", optional = true }
 transform-stream = "0.1.2"
 uuid = { version = "0.8.2", features = ["v4"] }
 xml-rs = "0.8.3"
+
+[dev-dependencies]
+tokio = { version = "1.9.0", features = ["full"] }
